remove remaining language picker code
