+++ frontends/main/src/app-pages/DashboardPage/CoursewareDisplay/model/dashboardViewModel.test.ts
@@ -16,7 +16,6 @@ import {
   buildVariantLabel,
   enrollmentCourseIsInPrograms,
   getCollectionFirstCoursesInDisplayOrder,
-  getDistinctDashboardLanguageOptions,
   getModuleCourseIdsFromPrograms,
   getNonContractProgramEnrollments,
   getProgramCoursesInContractOrder,
@@ -564,133 +563,6 @@ describe("dashboardViewModel", () => {
           programEnrollments: [],
         }),
       ).toEqual({ cards: [], initiallyVisibleCount: 0 })
-    })
-  })
-
-  describe("getDistinctDashboardLanguageOptions", () => {
-    test("includes enrollment language not present in enrollable language options", () => {
-      const englishRun = factories.courses.courseRun({
-        id: 101,
-        language: LanguageEnum.En,
-        courseware_id: "cw-en-101",
-        courseware_url: "https://example.com/en-101",
-        is_enrollable: true,
-      })
-
-      const course = factories.courses.course({
-        id: 1,
-        courseruns: [englishRun],
-        next_run_id: englishRun.id,
-        language_options: [
-          {
-            id: englishRun.id,
-            courseware_id: englishRun.courseware_id,
-            courseware_url: englishRun.courseware_url ?? "",
-            language: LanguageEnum.En,
-            title: englishRun.title,
-            run_tag: englishRun.run_tag,
-          },
-        ],
-      })
-
-      const esEnrollment = factories.enrollment.courseEnrollment({
-        run: {
-          id: 999,
-          language: LanguageEnum.EsEs,
-          course: { id: course.id },
-        },
-      })
-
-      const options = getDistinctDashboardLanguageOptions(
-        [course],
-        [esEnrollment],
-      )
-
-      expect(options.map((option) => option.value)).toEqual([
-        "language:en",
-        "language:es-es",
-      ])
-    })
-
-    test("adds only relevant enrollment languages for the provided course set", () => {
-      const courseA = factories.courses.course({ id: 10, language_options: [] })
-      const courseB = factories.courses.course({ id: 20, language_options: [] })
-
-      const enrollmentA = factories.enrollment.courseEnrollment({
-        run: { language: "fr", course: { id: 10 } },
-      })
-      const enrollmentOtherCourse = factories.enrollment.courseEnrollment({
-        run: { language: "de", course: { id: 999 } },
-      })
-
-      const options = getDistinctDashboardLanguageOptions(
-        [courseA, courseB],
-        [enrollmentA, enrollmentOtherCourse],
-      )
-
-      expect(options.map((option) => option.value)).toEqual(["language:fr"])
-    })
-
-    test("sorts enrollment-derived language options by label", () => {
-      const course = factories.courses.course({
-        id: 50,
-        language_options: [],
-        courseruns: [],
-      })
-      const frenchEnrollment = factories.enrollment.courseEnrollment({
-        run: { language: "fr", course: { id: 50 } },
-      })
-      const spanishEnrollment = factories.enrollment.courseEnrollment({
-        run: { language: LanguageEnum.EsEs, course: { id: 50 } },
-      })
-
-      const options = getDistinctDashboardLanguageOptions(
-        [course],
-        [frenchEnrollment, spanishEnrollment],
-      )
-
-      expect(options.map((option) => option.value)).toEqual([
-        "language:es-es",
-        "language:fr",
-      ])
-    })
-
-    test("uses the shared native language fallback label for enrollment-derived options", () => {
-      const originalDisplayNames = Intl.DisplayNames
-      Object.defineProperty(Intl, "DisplayNames", {
-        value: undefined,
-        configurable: true,
-        writable: true,
-      })
-
-      try {
-        const course = factories.courses.course({
-          id: 40,
-          language_options: [],
-          courseruns: [],
-        })
-        const enrollment = factories.enrollment.courseEnrollment({
-          run: { language: LanguageEnum.EsEs, course: { id: 40 } },
-        })
-
-        const options = getDistinctDashboardLanguageOptions(
-          [course],
-          [enrollment],
-        )
-
-        expect(options).toEqual([
-          {
-            value: "language:es-es",
-            label: "español",
-          },
-        ])
-      } finally {
-        Object.defineProperty(Intl, "DisplayNames", {
-          value: originalDisplayNames,
-          configurable: true,
-          writable: true,
-        })
-      }
     })
   })
